+++ affinda-api-client/src/main/java/com/affinda/api/client/models/ResumeDataPhoneNumberDetailsItem.java
@@ -1,0 +1,112 @@
+package com.affinda.api.client.models;
+
+import com.azure.core.annotation.Fluent;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+/** The ResumeDataPhoneNumberDetailsItem model. */
+@Fluent
+public final class ResumeDataPhoneNumberDetailsItem {
+    /*
+     * The rawText property.
+     */
+    @JsonProperty(value = "rawText")
+    private String rawText;
+
+    /*
+     * The formattedNumber property.
+     */
+    @JsonProperty(value = "formattedNumber")
+    private String formattedNumber;
+
+    /*
+     * The countryCode property.
+     */
+    @JsonProperty(value = "countryCode")
+    private String countryCode;
+
+    /*
+     * The nationalNumber property.
+     */
+    @JsonProperty(value = "nationalNumber")
+    private String nationalNumber;
+
+    /**
+     * Get the rawText property: The rawText property.
+     *
+     * @return the rawText value.
+     */
+    public String getRawText() {
+        return this.rawText;
+    }
+
+    /**
+     * Set the rawText property: The rawText property.
+     *
+     * @param rawText the rawText value to set.
+     * @return the ResumeDataPhoneNumberDetailsItem object itself.
+     */
+    public ResumeDataPhoneNumberDetailsItem setRawText(String rawText) {
+        this.rawText = rawText;
+        return this;
+    }
+
+    /**
+     * Get the formattedNumber property: The formattedNumber property.
+     *
+     * @return the formattedNumber value.
+     */
+    public String getFormattedNumber() {
+        return this.formattedNumber;
+    }
+
+    /**
+     * Set the formattedNumber property: The formattedNumber property.
+     *
+     * @param formattedNumber the formattedNumber value to set.
+     * @return the ResumeDataPhoneNumberDetailsItem object itself.
+     */
+    public ResumeDataPhoneNumberDetailsItem setFormattedNumber(String formattedNumber) {
+        this.formattedNumber = formattedNumber;
+        return this;
+    }
+
+    /**
+     * Get the countryCode property: The countryCode property.
+     *
+     * @return the countryCode value.
+     */
+    public String getCountryCode() {
+        return this.countryCode;
+    }
+
+    /**
+     * Set the countryCode property: The countryCode property.
+     *
+     * @param countryCode the countryCode value to set.
+     * @return the ResumeDataPhoneNumberDetailsItem object itself.
+     */
+    public ResumeDataPhoneNumberDetailsItem setCountryCode(String countryCode) {
+        this.countryCode = countryCode;
+        return this;
+    }
+
+    /**
+     * Get the nationalNumber property: The nationalNumber property.
+     *
+     * @return the nationalNumber value.
+     */
+    public String getNationalNumber() {
+        return this.nationalNumber;
+    }
+
+    /**
+     * Set the nationalNumber property: The nationalNumber property.
+     *
+     * @param nationalNumber the nationalNumber value to set.
+     * @return the ResumeDataPhoneNumberDetailsItem object itself.
+     */
+    public ResumeDataPhoneNumberDetailsItem setNationalNumber(String nationalNumber) {
+        this.nationalNumber = nationalNumber;
+        return this;
+    }
+}